feat: Add authentication, hrworks base class and fetching Onboarding Documents
For authentication this adds a custom got instance that will automatically take care of setting the authorization header.
This header has to be refreshed after 15 minutes.
The api is capable of getting the onboarding docs. Based on the status it needs to determine
if it has to create a user in Office365 or not.